Simplify tr/excpet and add an if

+++ src/python/TaskWorker/Actions/StageoutCheck.py
@@ -27,7 +27,8 @@ class StageoutCheck(TaskAction):
     def checkPermissions(self, Cmd):
         """
         Execute command and in case of permanent issue, raise error
-        If issue unknown, upload warning message
+        If issue unknown, upload warning message and return 1
+        Return 0 otherwise
         """
         self.logger.info("Executing command: %s " % Cmd)
         out, err, exitcode = executeCommand(Cmd)
@@ -44,7 +45,8 @@ class StageoutCheck(TaskAction):
                 msg = "The CRAB3 server got a non-critical error while checking stageout permissions. Please use checkwrite to check if everything is fine."
                 self.uploadWarning(msg, self.task['user_proxy'], self.task['tm_taskname'])
                 self.logger.info("UNKNOWN ERROR. Operator should check if it is permanent, but for now we go ahead and submit a task.")
-                return
+                return 1
+        return 0
 
 
     def execute(self, *args, **kw):
@@ -81,12 +83,10 @@ class StageoutCheck(TaskAction):
             createDummyFile(filename, self.logger)
             try:
                 self.logger.info("Executing cp command: %s " % cpCmd)
-                self.checkPermissions(cpCmd)
-            except:
-                raise
-            else:
-                self.logger.info("Executing rm command: %s " % rmCmd)
-                self.checkPermissions(rmCmd)
+                res = self.checkPermissions(cpCmd)
+                if res==0:
+                    self.logger.info("Executing rm command: %s " % rmCmd)
+                    self.checkPermissions(rmCmd)
             finally:
                 removeDummyFile(filename, self.logger)
         except IOError as er:
